Restructure c_src tree
c_src/
├── device/
│   ├── cairo/
│   │   ├── cairo_ctx.h
│   │   ├── cairo_fb.c
│   │   ├── cairo_font_ops.c
│   │   ├── cairo_image_ops.c
│   │   └── cairo_script_ops.c
│   ├── device.h
│   └── nvg/
│       ├── bcm.c
│       ├── drm.c
│       ├── glfw.c
│       ├── gl_helpers.c
│       ├── nanovg/
│       │   ├── LICENSE.txt
│       │   ├── nanovg.c
│       │   ├── nanovg_gl.h
│       │   ├── nanovg_gl_utils.h
│       │   └── nanovg.h
│       ├── nvg_font_ops.c
│       ├── nvg_image_ops.c
│       └── nvg_script_ops.c
├── font/
│   ├── font.c
│   ├── font.h
│   ├── font_ops.h
│   ├── fontstash.h
│   └── stb_truetype.h
├── image/
│   ├── image.c
│   ├── image.h
│   ├── image_ops.h
│   └── stb_image.h
├── main.c
├── scenic/
│   ├── common.h
│   ├── comms.c
│   ├── comms.h
│   ├── scenic_ops.c
│   ├── scenic_ops.h
│   ├── scenic_types.h
│   ├── script.c
│   ├── script.h
│   ├── script_ops.c
│   ├── script_ops.h
│   ├── unix_comms.c
│   ├── unix_comms.h
│   ├── utils.c
│   └── utils.h
└── tommyds/
