+++ Makefile
@@ -50,11 +50,23 @@ TOMMYDS_SRCS = \
 
 SCENIC_SRCS = \
 	c_src/scenic/comms.c \
-	c_src/scenic/ops/scenic_ops.c \
-	c_src/scenic/ops/script_ops.c \
+	c_src/scenic/scenic_ops.c \
+	c_src/scenic/script_ops.c \
 	c_src/scenic/script.c \
 	c_src/scenic/unix_comms.c \
 	c_src/scenic/utils.c
+
+NVG_COMMON_SRCS = \
+	c_src/device/nvg/gl_helpers.c \
+	c_src/device/nvg/nanovg/nanovg.c \
+	c_src/device/nvg/nvg_font_ops.c \
+	c_src/device/nvg/nvg_image_ops.c \
+	c_src/device/nvg/nvg_script_ops.c
+
+CAIRO_COMMON_SRCS = \
+	c_src/device/cairo/cairo_font_ops.c \
+	c_src/device/cairo/cairo_image_ops.c \
+	c_src/device/cairo/cairo_script_ops.c
 
 ifeq ($(SCENIC_LOCAL_TARGET),glfw)
 	CFLAGS = -O3 -std=c99
@@ -85,12 +97,8 @@ ifeq ($(SCENIC_LOCAL_TARGET),glfw)
 	endif
 
 	DEVICE_SRCS += \
-		c_src/device/glfw.c \
-		c_src/device/gl_helpers.c \
-		c_src/nanovg/nanovg.c
-	FONT_SRCS += c_src/font/nvg_font_ops.c
-	IMAGE_SRCS += c_src/image/nvg_image_ops.c
-	SCENIC_SRCS += c_src/scenic/ops/nvg_script_ops.c
+		$(NVG_COMMON_SRCS) \
+		c_src/device/nvg/glfw.c
 
 else ifeq ($(SCENIC_LOCAL_TARGET),bcm)
 	LDFLAGS += -lGLESv2 -lEGL -lm -lvchostif -lbcm_host
@@ -98,12 +106,8 @@ else ifeq ($(SCENIC_LOCAL_TARGET),bcm)
 	CFLAGS += -std=gnu99
 
 	DEVICE_SRCS += \
-		c_src/device/bcm.c \
-		c_src/device/gl_helpers.c \
-		c_src/nanovg/nanovg.c
-	FONT_SRCS += c_src/font/nvg_font_ops.c
-	IMAGE_SRCS += c_src/image/nvg_image_ops.c
-	SCENIC_SRCS += c_src/scenic/ops/nvg_script_ops.c
+		$(NVG_COMMON_SRCS) \
+		c_src/device/nvg/bcm.c
 
 	ifeq ($(SCENIC_LOCAL_GL),gles2)
 		CFLAGS += -DSCENIC_GLES2
@@ -119,29 +123,26 @@ else ifeq ($(SCENIC_LOCAL_TARGET),drm)
 	CFLAGS += -fPIC -I$(NERVES_SDK_SYSROOT)/usr/include/drm
 
 	DEVICE_SRCS += \
-		c_src/device/drm.c \
-		c_src/device/gl_helpers.c \
-		c_src/nanovg/nanovg.c
-	FONT_SRCS += c_src/font/nvg_font_ops.c
-	IMAGE_SRCS += c_src/image/nvg_image_ops.c
-	SCENIC_SRCS += c_src/scenic/ops/nvg_script_ops.c
+		$(NVG_COMMON_SRCS) \
+		c_src/device/nvg/drm.c
 
 	ifeq ($(SCENIC_LOCAL_GL),gles2)
 		CFLAGS += -DSCENIC_GLES2
 	else
 		CFLAGS += -DSCENIC_GLES3
 	endif
-else ifeq ($(SCENIC_LOCAL_TARGET),cairo)
+
+else ifeq ($(SCENIC_LOCAL_TARGET),cairo-fb)
 	LDFLAGS += `pkg-config --static --libs freetype2 cairo`
 	CFLAGS += `pkg-config --static --cflags freetype2 cairo`
 	LDFLAGS += -lm
 	CFLAGS ?= -O2 -Wall -Wextra -Wno-unused-parameter -pedantic
 	CFLAGS += -std=gnu99
 
-	DEVICE_SRCS += c_src/device/cairo.c
-	FONT_SRCS += c_src/font/cairo_font_ops.c
-	IMAGE_SRCS += c_src/image/cairo_image_ops.c
-	SCENIC_SRCS += c_src/scenic/ops/cairo_script_ops.c
+	DEVICE_SRCS += \
+		$(CAIRO_COMMON_SRCS) \
+		c_src/device/cairo/cairo_fb.c
+
 else
 $(info ------ no SCENIC_LOCAL_TARGET set ------)
 $(info If you get here, then you are probably using a custom Nerves system)
@@ -173,7 +174,6 @@ CFLAGS += \
 	-Ic_src/device \
 	-Ic_src/font \
 	-Ic_src/image \
-	-Ic_src/nanovg \
 	-Ic_src/scenic \
 	-Ic_src/tommyds/src
