+++ c_src/device/cairo/cairo_image_ops.c
