+++ c_src/scenic/scenic_ops.h
