+++ c_src/scenic/script_ops.c
